Rename file with close codes
