+++ src/shared/type/websocket-close-code.ts
